requirements.txt
altgraph==0.17.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
comtypes==1.1.14
cryptography==37.0.4
cycler==0.11.0
fonttools==4.35.0
future==0.18.2
httplib2==0.20.4
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
oauth2==1.9.0.post1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pefile==2022.5.30
Pillow==9.2.0
Pint==0.19.2
playsound==1.3.0
PyAudio==0.2.12
pycparser==2.21
pygame==2.1.2
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.2.1
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.1
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
unit-converter==1.1.0
urllib3==1.26.10
